server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Werkzeug
bcrypt
psycopg2-binary
python-dotenv
sqlalchemy-serializer
Flask-JWT-Extended
